added authentication and authorization functionality

+++ controllers/coursesCOntroller.js
@@ -42,11 +42,24 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access   private
 exports.createCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  // add user to req.body
+  req.body.user = req.user.id;
+
   const bootcamp = await Bootcomp.findById(req.body.bootcamp);
   if (!bootcamp)
     return next(
       new errorResponse(`No Bootcamp found with id ${req.params.id}`, 404)
     );
+
+  // Ensure that the user owns this bootcamp
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new errorResponse(
+        `You are not allowed to add acourse to this bootcamp`,
+        403
+      )
+    );
+  }
 
   const data = await Courses.create(req.body);
   res.status(201).send({
@@ -60,7 +73,15 @@ exports.createCourse = asyncHandler(async (req, res, next) => {
 // @route   /api/vi/courses/id
 // @access   private
 exports.updateCourse = asyncHandler(async (req, res, next) => {
-  const data = await Courses.findByIdAndUpdate(req.params.id, req.body, {
+  let data = await Courses.findById(req.params.id);
+
+  // Ensure that the user owns this bootcamp
+  if (data.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new errorResponse(`You are not allowed to edit this course `, 403)
+    );
+  }
+  data = await Courses.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true,
   });
@@ -78,11 +99,21 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 // @route   /api/vi/courses/id
 // @access   private
 exports.deleteCourse = asyncHandler(async (req, res, next) => {
-  const data = await Courses.findByIdAndRemove(req.params.id);
+  let data = await Courses.findById(req.params.id);
+
+  // Ensure that the user owns this bootcamp
+  if (data.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new errorResponse(`You are not allowed to delete this course `, 403)
+    );
+  }
+
+  data = await Courses.findByIdAndRemove(req.params.id);
   if (!data)
     return next(
       new errorResponse(`No Course found with id ${req.params.id}`, 404)
     );
+
   res.status(200).send({
     status: "success",
     data: {},
